Avatar.tsx GearList.tsx ItemRow.tsx separated
