+++ client/hooks/itemrow/useItemRow.ts
@@ -1,0 +1,9 @@
+import { useState } from 'react'
+
+export const useItemRow = () => {
+    const [isChecked, setChecked] = useState(false);
+    const handleChange = (value: boolean) => {
+        setChecked(value);
+    }
+    return { isChecked, handleChange };
+}
